+++ src/main/java/jp/mydns/projectk/safi/dxo/JobDxo.java
@@ -34,9 +34,9 @@ import java.util.Optional;
 import jp.mydns.projectk.safi.constant.JobStatus;
 import jp.mydns.projectk.safi.entity.JobEntity;
 import jp.mydns.projectk.safi.service.JsonService;
+import static jp.mydns.projectk.safi.util.CollectionUtils.convertElements;
 import jp.mydns.projectk.safi.util.TimeUtils;
 import jp.mydns.projectk.safi.util.JsonValueUtils;
-import static jp.mydns.projectk.safi.util.LambdaUtils.convertElements;
 import jp.mydns.projectk.safi.value.JobCreationContext;
 import jp.mydns.projectk.safi.value.JobValue;
 import jp.mydns.projectk.safi.value.JobdefValue;
